+++ src/Pipe.php
@@ -9,37 +9,107 @@ declare(strict_types=1);
 
 namespace Serafim\Pipe;
 
-use Serafim\Pipe\Exception\FunctionNotFoundException;
 use Serafim\Pipe\Resolver\AsIs;
 use Serafim\Pipe\Resolver\Join;
-use Serafim\Pipe\Resolver\ResolverInterface;
 use Serafim\Pipe\Resolver\Snake;
 use Serafim\Placeholder\Placeholder;
+use Serafim\Pipe\Resolver\ResolverInterface;
+use Serafim\Pipe\Exception\FunctionNotFoundException;
 
 /**
- * Class Pipe
+ * Object-oriented pipe operator implementation based on PHP Pipe operator.
  *
+ * A common PHP OOP pattern is the use of method chaining, or what is
+ * also known as "Fluent Expressions". So named for the way one method
+ * flows into the next to form a conceptual hyper-expression.
+ *
+ * However, when using the functional approach this can lead to reduced
+ * readability, polluted symbol tables, or static-analysis defying
+ * type inconsistency such as in the following example:
+ *
+ * <code>
+ *  $snakeCase = strtolower(
+ *       preg_replace('/(.)(?=[A-Z])/u', '$1_',
+ *           preg_replace('/\s+/u', '',
+ *               ucwords('HelloWorld')
+ *           )
+ *       )
+ *  );
+ *
+ *  var_dump($snakeCase);
+ *  // expected output: "hello_world"
+ * </code>
+ *
+ * The pipe fixes this problem, allows you to
+ * chain the execution of pure functions:
+ *
+ * <code>
+ *
+ *  $snakeCase = pipe($camelCase)
+ *      ->ucWords(_)
+ *      ->pregReplace('/\s+/u', '', _)
+ *      ->pregReplace('/(.)(?=[A-Z])/u', '$1_', _)
+ *      ->strToLower(_)
+ *      ->varDump;
+ *  // expected output: "hello_world"
+ *
+ * </code>
+ *
+ * @see https://wiki.php.net/rfc/pipe-operator
  * @noinspection PhpUndefinedClassInspection
- * @property $value
  */
-final class Pipe
+final class Pipe implements PipeInterface
 {
-    /**
-     * @var string
-     */
-    private const VALUE_PROPERTY_SELECTOR = 'value';
+    use RendererTrait;
+    use OperatorsTrait;
 
     /**
+     * A constant that defines the format of the error messages
+     * to find a suitable function.
+     *
+     * @var string
+     */
+    private const ERROR_NOT_FOUND = 'Function "%s" not found';
+
+    /**
+     * @var string[]|ResolverInterface[]
+     */
+    private const DEFAULT_RESOLVERS = [
+        AsIs::class,
+        Snake::class,
+        Join::class,
+    ];
+
+    /**
+     * A set of keys for functions for quicker recall in format:
+     * <code>
+     *  [
+     *      'vardump' => 'var_dump',
+     *      'strreplace' => 'str_replace'
+     *  ]
+     * </code>
+     *
+     * @var array|string[]
+     */
+    private static $memoized = [];
+
+    /**
+     * Original value with which we interact.
+     *
+     * @var mixed
+     */
+    protected $value;
+
+    /**
+     * List of objects that define a set of renaming algorithms.
+     *
      * @var array|ResolverInterface[]
      */
     private $resolvers;
 
     /**
-     * @var mixed
-     */
-    private $value;
-
-    /**
+     * The namespace that we are currently using for function prefixes.
+     *
      * @var string
      */
     private $namespace = '';
@@ -49,22 +119,41 @@ final class Pipe
      *
      * @param mixed $value
      */
-    public function __construct($value)
+    public function __construct($value = null)
     {
-        $this->value = $value;
+        $this->value = $this->resolveValue($value);
         $this->resolvers = $this->bootDefaultResolvers();
     }
 
     /**
+     * Method to normalize the value to the valid for the job.
+     *
+     * @param mixed $value
+     * @return mixed
+     */
+    private function resolveValue($value)
+    {
+        if ($value instanceof self) {
+            return $value();
+        }
+
+        return $value;
+    }
+
+    /**
+     * Constructor for default function name resolvers.
+     *
      * @return array|ResolverInterface[]
      */
     private function bootDefaultResolvers(): array
     {
-        return [
-            new AsIs(),
-            new Snake(),
-            new Join(),
-        ];
+        $result = [];
+
+        foreach (self::DEFAULT_RESOLVERS as $class) {
+            $result[] = new $class();
+        }
+
+        return $result;
     }
 
     /**
@@ -72,7 +161,9 @@ final class Pipe
      */
     public function __debugInfo(): array
     {
-        return [self::VALUE_PROPERTY_SELECTOR => $this->value];
+        return [
+            'value' => $this->value,
+        ];
     }
 
     /**
@@ -89,6 +180,14 @@ final class Pipe
     }
 
     /**
+     * Method to replace placeholders in the argument list
+     * with the original value.
+     *
+     * <code>
+     *  $this->applyArguments([1, _, _]);
+     *  // expected output: [1, $this->value, $this->value]
+     * </code>
+     *
      * @param array $arguments
      * @return array
      */
@@ -100,10 +199,13 @@ final class Pipe
     }
 
     /**
+     * Method to specify the current namespace with which
+     * we will work further.
+     *
      * @param string $namespace
-     * @return Pipe|$this
+     * @return PipeInterface|Pipe|$this
      */
-    public function use(string $namespace): self
+    public function use(string $namespace): PipeInterface
     {
         $self = clone $this;
 
@@ -113,40 +215,71 @@ final class Pipe
     }
 
     /**
-     * @param string $name
+     * @param string $function
      * @param array $arguments
-     * @return Pipe|$this
+     * @return PipeInterface|Pipe|$this
      */
-    public function __call(string $name, array $arguments = []): self
+    public function __call(string $function, array $arguments = []): PipeInterface
     {
-        return $this->exec($name, $this->applyArguments($arguments));
+        return $this->exec($function, $this->applyArguments($arguments));
     }
 
     /**
+     * Method to call the desired function.
+     *
      * @param string $function
      * @param array $arguments
      * @return Pipe|$this
      */
     private function exec(string $function, array $arguments): self
     {
-        $function = $this->lookup($function, function (string $name): ?string {
-            $function = $this->namespaced($name);
-
-            return \function_exists($function) ? $function : null;
-        });
+        if (($normalized = $this->resolveMemoizedName($function)) === null) {
+            throw $this->functionNotFound($function);
+        }
 
         $self = clone $this;
-        $self->value = $function(...$arguments);
+        $self->value = $normalized(...$arguments);
 
         return $self;
     }
 
     /**
+     * Getting the normal name of the function that we should call in the
+     * future with the intermediate memoization of the original name.
+     *
+     * @param string $function
+     * @return string|null
+     */
+    private function resolveMemoizedName(string $function): ?string
+    {
+        $lower = \strtolower($function);
+
+        return self::$memoized[$lower] ?? self::$memoized[$lower] = $this->resolveName($function);
+    }
+
+    /**
+     * Getting the normal name of the function that we should call in the future.
+     *
+     * @param string $function
+     * @return string|null
+     */
+    private function resolveName(string $function): ?string
+    {
+        return $this->lookup($function, function (string $name): ?string {
+            $function = $this->namespaced($name);
+
+            return \function_exists($function) ? $function : null;
+        });
+    }
+
+    /**
+     * Lookup method to a suitable function name.
+     *
      * @param string $function
      * @param \Closure $then
-     * @return string
+     * @return string|null
      */
-    private function lookup(string $function, \Closure $then): string
+    private function lookup(string $function, \Closure $then): ?string
     {
         foreach ($this->resolvers as $resolver) {
             if (\is_string($name = $then($resolver->resolve($function)))) {
@@ -154,11 +287,12 @@ final class Pipe
             }
         }
 
-        $error = \sprintf('Function "%s" not found', $this->namespaced($function));
-        throw new FunctionNotFoundException($error);
+        return null;
     }
 
     /**
+     * Method that adds a namespace prefix to the function being called.
+     *
      * @param string $name
      * @return string
      */
@@ -168,16 +302,56 @@ final class Pipe
     }
 
     /**
-     * @noinspection MagicMethodsValidityInspection
-     * @param string $name
-     * @return mixed|$this
+     * @param string $function
+     * @return \Serafim\Pipe\Exception\FunctionNotFoundException
      */
-    public function __get(string $name)
+    private function functionNotFound(string $function): FunctionNotFoundException
     {
-        if (\strtolower($name) === 'value') {
-            return $this->value;
+        $message = \sprintf(self::ERROR_NOT_FOUND, $this->namespaced($function));
+
+        return new FunctionNotFoundException($message);
+    }
+
+    /**
+     * @noinspection MagicMethodsValidityInspection
+     *
+     * @param string $function
+     * @return \Serafim\Pipe\PipeInterface|Pipe|$this
+     */
+    public function __get(string $function): PipeInterface
+    {
+        if (\method_exists(OperatorsTrait::class, $function)) {
+            return $this->$function($this->value);
         }
 
-        return $this->exec($name, [$this->value]);
+        return $this->exec($function, [$this->value]);
+    }
+
+    /**
+     * @noinspection MagicMethodsValidityInspection
+     *
+     * @param string $function
+     * @return bool
+     */
+    public function __isset(string $function): bool
+    {
+        return $this->resolveMemoizedName($function) !== null;
+    }
+
+    /**
+     * @return string
+     */
+    public function __toString(): string
+    {
+        return $this->render($this->value);
+    }
+
+    /**
+     * @param mixed $value
+     * @return mixed
+     */
+    private function applyArgument($value)
+    {
+        return Placeholder::match($value) ? $this->value : $value;
     }
 }